+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -126,7 +126,7 @@ public class CategoryServiceImpl implements CategoryService {
      * @param type
      * @return
      */
-    public List<Category> list(Long type) {
+    public List<Category> list(Integer type) {
         return categoryMapper.list(type);
     }
 }
